// types.ts
export type MapItemType = 'POI' | 'BUILDING' | 'ROAD' | 'OBSTACLE' | 'CHEST' | 'TRIGGER' | 'FAMILY' | 'PLAYER' | 'BOSS' | 'OTHER';

export interface MapItem {
    id: string;
    x: number;
    z: number;
    type: MapItemType;
    label?: string;
    icon?: string; // Emoji or SVG path
    color?: string;
    radius?: number; // Size on map
}